+++ docs/court_wizard.js
@@ -256,24 +256,24 @@ if (!('encodeInto' in cachedTextEncoder)) {
 
 let WASM_VECTOR_LEN = 0;
 
-function __wasm_bindgen_func_elem_127765(arg0, arg1, arg2) {
-    wasm.__wasm_bindgen_func_elem_127765(arg0, arg1, addHeapObject(arg2));
+function __wasm_bindgen_func_elem_128069(arg0, arg1, arg2) {
+    wasm.__wasm_bindgen_func_elem_128069(arg0, arg1, addHeapObject(arg2));
 }
 
-function __wasm_bindgen_func_elem_126205(arg0, arg1, arg2) {
-    wasm.__wasm_bindgen_func_elem_126205(arg0, arg1, addHeapObject(arg2));
+function __wasm_bindgen_func_elem_123537(arg0, arg1) {
+    wasm.__wasm_bindgen_func_elem_123537(arg0, arg1);
 }
 
-function __wasm_bindgen_func_elem_127764(arg0, arg1) {
-    wasm.__wasm_bindgen_func_elem_127764(arg0, arg1);
+function __wasm_bindgen_func_elem_128070(arg0, arg1, arg2, arg3) {
+    wasm.__wasm_bindgen_func_elem_128070(arg0, arg1, addHeapObject(arg2), addHeapObject(arg3));
 }
 
-function __wasm_bindgen_func_elem_123233(arg0, arg1) {
-    wasm.__wasm_bindgen_func_elem_123233(arg0, arg1);
+function __wasm_bindgen_func_elem_128068(arg0, arg1) {
+    wasm.__wasm_bindgen_func_elem_128068(arg0, arg1);
 }
 
-function __wasm_bindgen_func_elem_127766(arg0, arg1, arg2, arg3) {
-    wasm.__wasm_bindgen_func_elem_127766(arg0, arg1, addHeapObject(arg2), addHeapObject(arg3));
+function __wasm_bindgen_func_elem_126509(arg0, arg1, arg2) {
+    wasm.__wasm_bindgen_func_elem_126509(arg0, arg1, addHeapObject(arg2));
 }
 
 const __wbindgen_enum_GpuAddressMode = ["clamp-to-edge", "repeat", "mirror-repeat"];
@@ -2222,14 +2222,9 @@ function __wbg_get_imports() {
         const ret = getObject(arg0).y;
         return ret;
     };
-    imports.wbg.__wbindgen_cast_07f5d3f4187e36d3 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 84354, function: Function { arguments: [NamedExternref("FocusEvent")], shim_idx: 84355, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127643, __wasm_bindgen_func_elem_127765);
-        return addHeapObject(ret);
-    };
-    imports.wbg.__wbindgen_cast_1e1d2f516325a803 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 84354, function: Function { arguments: [NamedExternref("Array<any>")], shim_idx: 84355, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127643, __wasm_bindgen_func_elem_127765);
+    imports.wbg.__wbindgen_cast_1e0ce76af689e38e = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 84604, function: Function { arguments: [NamedExternref("WheelEvent")], shim_idx: 84605, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127947, __wasm_bindgen_func_elem_128069);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_cast_2241b6af4c4b2941 = function(arg0, arg1) {
@@ -2237,39 +2232,44 @@ function __wbg_get_imports() {
         const ret = getStringFromWasm0(arg0, arg1);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_27cc381950cca288 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 84354, function: Function { arguments: [NamedExternref("PointerEvent")], shim_idx: 84355, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127643, __wasm_bindgen_func_elem_127765);
+    imports.wbg.__wbindgen_cast_281dab70b4069fd3 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 84604, function: Function { arguments: [NamedExternref("Array<any>")], shim_idx: 84605, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127947, __wasm_bindgen_func_elem_128069);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_30e88d20283b7a27 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 84354, function: Function { arguments: [], shim_idx: 84356, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127643, __wasm_bindgen_func_elem_127764);
+    imports.wbg.__wbindgen_cast_4008fcd2a137a8d0 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 84546, function: Function { arguments: [Externref], shim_idx: 84547, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_126499, __wasm_bindgen_func_elem_126509);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_5b3bc3af97c85c87 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 84354, function: Function { arguments: [NamedExternref("Event")], shim_idx: 84355, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127643, __wasm_bindgen_func_elem_127765);
+    imports.wbg.__wbindgen_cast_61c56cc19ccc1343 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 84604, function: Function { arguments: [], shim_idx: 84606, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127947, __wasm_bindgen_func_elem_128068);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_5d61dab7e66a19a7 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 84354, function: Function { arguments: [NamedExternref("WheelEvent")], shim_idx: 84355, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127643, __wasm_bindgen_func_elem_127765);
+    imports.wbg.__wbindgen_cast_63e626533d168060 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 84604, function: Function { arguments: [NamedExternref("Event")], shim_idx: 84605, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127947, __wasm_bindgen_func_elem_128069);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_8eb744c80a93002b = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 83728, function: Function { arguments: [], shim_idx: 83729, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_123212, __wasm_bindgen_func_elem_123233);
+    imports.wbg.__wbindgen_cast_6b98a133d2b84814 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 84604, function: Function { arguments: [NamedExternref("PointerEvent")], shim_idx: 84605, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127947, __wasm_bindgen_func_elem_128069);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_94a4264746fdcb03 = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 84296, function: Function { arguments: [Externref], shim_idx: 84297, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_126195, __wasm_bindgen_func_elem_126205);
+    imports.wbg.__wbindgen_cast_91862625ed0b6379 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 84604, function: Function { arguments: [NamedExternref("PageTransitionEvent")], shim_idx: 84605, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127947, __wasm_bindgen_func_elem_128069);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_b7a9bf1c1a875fcc = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 84354, function: Function { arguments: [NamedExternref("Array<any>"), NamedExternref("ResizeObserver")], shim_idx: 84357, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127643, __wasm_bindgen_func_elem_127766);
+    imports.wbg.__wbindgen_cast_a24c05f99abd5fde = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 84604, function: Function { arguments: [NamedExternref("FocusEvent")], shim_idx: 84605, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127947, __wasm_bindgen_func_elem_128069);
+        return addHeapObject(ret);
+    };
+    imports.wbg.__wbindgen_cast_bd91b667fc2cede6 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 83978, function: Function { arguments: [], shim_idx: 83979, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_123516, __wasm_bindgen_func_elem_123537);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_cast_cb9088102bce6b30 = function(arg0, arg1) {
@@ -2277,19 +2277,19 @@ function __wbg_get_imports() {
         const ret = getArrayU8FromWasm0(arg0, arg1);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_cast_cc305dc5871b94ea = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 84354, function: Function { arguments: [NamedExternref("KeyboardEvent")], shim_idx: 84355, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127643, __wasm_bindgen_func_elem_127765);
-        return addHeapObject(ret);
-    };
-    imports.wbg.__wbindgen_cast_d23c06fbd024eb4d = function(arg0, arg1) {
-        // Cast intrinsic for `Closure(Closure { dtor_idx: 84354, function: Function { arguments: [NamedExternref("PageTransitionEvent")], shim_idx: 84355, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
-        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127643, __wasm_bindgen_func_elem_127765);
-        return addHeapObject(ret);
-    };
     imports.wbg.__wbindgen_cast_d6cd19b81560fd6e = function(arg0) {
         // Cast intrinsic for `F64 -> Externref`.
         const ret = arg0;
+        return addHeapObject(ret);
+    };
+    imports.wbg.__wbindgen_cast_d8f31d1b7b74c567 = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 84604, function: Function { arguments: [NamedExternref("Array<any>"), NamedExternref("ResizeObserver")], shim_idx: 84607, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127947, __wasm_bindgen_func_elem_128070);
+        return addHeapObject(ret);
+    };
+    imports.wbg.__wbindgen_cast_fbc332aba3915d1c = function(arg0, arg1) {
+        // Cast intrinsic for `Closure(Closure { dtor_idx: 84604, function: Function { arguments: [NamedExternref("KeyboardEvent")], shim_idx: 84605, ret: Unit, inner_ret: Some(Unit) }, mutable: true }) -> Externref`.
+        const ret = makeMutClosure(arg0, arg1, wasm.__wasm_bindgen_func_elem_127947, __wasm_bindgen_func_elem_128069);
         return addHeapObject(ret);
     };
     imports.wbg.__wbindgen_object_clone_ref = function(arg0) {