+++ src/game/units/wizard/spells/black_hole/mod.rs
@@ -1,6 +1,6 @@
 //! Black Hole spell module.
 
-mod components;
+pub(in crate::game::units::wizard::spells) mod components;
 pub(in crate::game::units::wizard) mod constants;
 mod plugin;
 mod systems;