+++ src/game/units/wizard/spells/magic_missile/mod.rs
@@ -2,7 +2,7 @@
 //!
 //! Handles magic missile projectiles that home in on attackers.
 
-mod components;
+pub(in crate::game::units::wizard::spells) mod components;
 pub(in crate::game::units::wizard) mod constants;
 mod plugin;
 mod styles;